rename files to avoid dots
