+++ custom_components/webrtc/__init__.py
@@ -54,7 +54,7 @@ LINKS = {}  # 2 3 4
 async def async_setup(hass: HomeAssistantType, config: ConfigType):
     # 1. Serve lovelace card
     path = Path(__file__).parent / "www"
-    for name in ("video-rtc.js", "webrtc-camera.js", "digital-ptz.js", "ditigal-ptz.transform.js", "digital-ptz.gestures.js"):
+    for name in ("video-rtc.js", "webrtc-camera.js", "digital-ptz.js", "ditigal-ptz-transform.js", "digital-ptz-gestures.js"):
         utils.register_static_path(
             hass.http.app, "/webrtc/" + name, path / name)
