+++ custom_components/webrtc/www/digital-ptz.js
@@ -7,9 +7,8 @@ import {
   startOneFingerPan,
   startPinchZoom,
   startTouchTapDragZoom,
-} from "./digital-ptz.gestures.js";
-import { Transform } from "./ditigal-ptz.transform.js";
-
+} from "./digital-ptz-gestures.js";
+import { Transform } from "./ditigal-ptz-transform.js";
 export const ONE_FINGER_ZOOM_SPEED = 1 / 200; // 1 scale every 200px
 export const DBL_CLICK_MS = 400;
 export const MAX_ZOOM = 10;
